Make basic restructuring
