+++ application/services/web/src/app/(landing)/page.tsx
@@ -1,10 +1,10 @@
-import {classname} from '@/ui/common/utils/classname'
-import {Bullerin} from '@/ui/home/components/bulletin'
-import {Contribution} from '@/ui/home/components/contribution'
-import {ExploreWorld} from '@/ui/home/components/explore-world'
-import {LandingLayout} from '@/ui/home/components/landing-layout'
-import {Meetings} from '@/ui/home/components/meeting'
-import {Roadmap} from '@/ui/home/components/roadmap'
+import {Bullerin} from '@/common/home/components/bulletin'
+import {Contribution} from '@/common/home/components/contribution'
+import {ExploreWorld} from '@/common/home/components/explore-world'
+import {LandingLayout} from '@/common/home/components/landing-layout'
+import {Meetings} from '@/common/home/components/meeting'
+import {Roadmap} from '@/common/home/components/roadmap'
+import {classname} from '@/common/utils/classname'
 
 export default function Home() {
   return (
