+++ application/services/web/src/auth/components/join-form/lib/get-button-label.ts
@@ -1,0 +1,9 @@
+import {JoinFormVariantEnum} from '@/auth/components/join-form/join-form'
+
+export const getButtonLabel = (variant: JoinFormVariantEnum) => {
+  if (variant === JoinFormVariantEnum.login) {
+    return 'Login'
+  }
+
+  return 'Sign Up'
+}
