+++ application/services/web/src/auth/lib/token.server.ts
@@ -1,0 +1,5 @@
+'use server'
+
+import {cookies} from 'next/headers'
+
+export const getServerToken = () => cookies().get('accessToken')?.value ?? null
