+++ application/services/web/src/auth/requests/iam/iam.server.request.ts
@@ -1,0 +1,14 @@
+'use server'
+
+import {cookies} from 'next/headers'
+
+import {UserResponse} from '@/auth/interfaces/user-response.interface'
+import {requestFetchIam} from '@/auth/requests/iam/iam.request'
+import {getHeaders} from '@/common/lib/request/headers.server'
+
+export const iamServerRequest = async (): Promise<UserResponse | null> => {
+  const accessToken = cookies().get('accessToken')?.value
+  return accessToken
+    ? await requestFetchIam({accessToken, headers: getHeaders()})
+    : null
+}
