+++ application/services/web/src/auth/requests/login.request.ts
@@ -1,5 +1,5 @@
 import {TokenResponse} from '@/auth/interfaces/token-response.interface'
-import {PASSPORT_API_URL} from '@/auth/requests/constants'
+import {PASSPORT_API_URL} from '@/common/lib/request/constants'
 import {request} from '@/common/lib/request/request'
 
 export const requestLogin = async (data: any): Promise<TokenResponse> =>
