+++ application/services/web/src/auth/requests/register.request.ts
@@ -1,8 +1,7 @@
+import {PASSPORT_API_URL} from '@/common/lib/request/constants'
 import {request} from '@/common/lib/request/request'
 
 import {TokenResponse} from '../interfaces/token-response.interface'
-
-import {PASSPORT_API_URL} from './constants'
 
 export const requestRegister = async (data: any): Promise<TokenResponse> =>
   request<TokenResponse>(`${PASSPORT_API_URL}/authentication/register`, {
