+++ application/services/web/src/common/styles/globals.css
