+++ application/services/web/src/room/requests/create-room.request.ts
@@ -1,5 +1,5 @@
 import {TokenResponse} from '@/auth/interfaces/token-response.interface'
-import {TELEGRAPH_API_URL} from '@/auth/requests/constants'
+import {TELEGRAPH_API_URL} from '@/common/lib/request/constants'
 import {request} from '@/common/lib/request/request'
 import {CreateRoomResponse} from '@/room/interfaces/create-room-response.interface'
