--- application/services/web/src/ui/auth/components/lib/get-button-label.ts
@@ -1,9 +1,0 @@
-import {JoinFormVariantEnum} from '@/ui/auth/components/join-form'
-
-export const getButtonLabel = (variant: JoinFormVariantEnum) => {
-  if (variant === JoinFormVariantEnum.login) {
-    return 'Login'
-  }
-
-  return 'Sign Up'
-}
