--- application/services/web/src/ui/auth/lib/server-token.ts
@@ -1,3 +1,0 @@
-import {cookies} from 'next/headers'
-
-export const getServerToken = () => cookies().get('accessToken')?.value ?? null
