+++ application/services/web/tailwind.config.ts
@@ -1,14 +1,10 @@
 import type {Config} from 'tailwindcss'
 
 const config: Config = {
-  content: [
-    './src/ui/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
-  ],
+  content: ['./src/**/*.{js,ts,jsx,tsx,mdx}'],
   plugins: [],
   theme: {
-    extend: {
-    },
+    extend: {},
     fontFamily: {
       body: ['Inter', 'sans-serif'],
     },
